--- src/main/java/team/silvertown/masil/masil/dto/CreateResponse.java
@@ -1,5 +1,0 @@
-package team.silvertown.masil.masil.dto;
-
-public record CreateResponse(Long id) {
-
-}
