+++ src/main/java/team/silvertown/masil/masil/dto/request/CreateRequest.java
@@ -1,4 +1,4 @@
-package team.silvertown.masil.masil.dto;
+package team.silvertown.masil.masil.dto.request;
 
 
 import java.time.OffsetDateTime;
