+++ src/main/java/team/silvertown/masil/masil/dto/response/CreateResponse.java
@@ -1,0 +1,5 @@
+package team.silvertown.masil.masil.dto.response;
+
+public record CreateResponse(Long id) {
+
+}
